+++ rtx_core/src/token.rs
@@ -627,7 +627,7 @@ impl<'a> Token {
 
   pub fn to_register(&self, state: &State) -> Option<Rc<RefCell<Register>>> { state.lookup_register_definition(self) }
 
-  pub fn to_number(&self) -> Number { Number::new(self.text.parse::<i32>().unwrap_or(0)) }
+  pub fn to_number(&self) -> Number { Number::new(self.text.parse::<f32>().unwrap_or(0.0)) }
 
   pub fn value_of(&self, args: Vec<Token>, state: &mut State) -> Option<RegisterValue> {
     match self.to_register(state) {
